feat(virtual-dom): remove event listener module rename interfaces to types
