+++ package.json
@@ -62,6 +62,8 @@
   },
   "typings": "lib/es2015/index.d.ts",
   "dependencies": {
+    "@most/dom-event": "^1.3.2",
+    "@most/hold": "^1.4.2",
     "@most/prelude": "^1.4.1",
     "most": "^1.1.1"
   }
