+++ src/driver/index.ts
@@ -1,0 +1,1 @@
+export * from './makeDomDriver';
