+++ src/virtual-dom/index.ts
@@ -1,4 +1,4 @@
-export * from './interfaces';
+export * from './types';
 export * from './helpers/h';
 export * from './helpers/hyperscript';
 export * from './helpers/svg';
