+++ src/virtual-dom/modules/index.ts
@@ -1,7 +1,6 @@
 export * from './attributes';
 export * from './class';
 export * from './dataset';
-export * from './eventlisteners';
 export * from './hero';
 export * from './props';
 export * from './style';
