+++ src/virtual-dom/modules/style.ts
@@ -1,4 +1,4 @@
-import { VNode, Module } from '../interfaces';
+import { VNode, Module } from '../types';
 
 let requestAnimationFrame: any;
